# application/services/custom_services/catalogitem/src/models/curd.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const curdSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String,
   description: String,
   tag: String
})

const curdModel = mongoose.model('curd', curdSchema, 'curd');
export default curdModel;